requirements.txt
flask==3.0.0
flask-cors==3.0.10
python-dotenv==1.0.0
pymongo==4.10.1
dnspython==2.4.2
requests==2.32.3
PyJWT==2.9.0
dedalus-labs==0.1.0a9